included Slim, added print CSS
