+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "TPSBIN/Slim"]
+	path = TPSBIN/Slim
+	url = https://github.com/slimphp/Slim
