+++ Playlist/API/Library.php
@@ -1,0 +1,18 @@
+<?php
+
+/* 
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+class LibraryAPI {
+
+    function __construct() {
+        
+    }
+    function getList(){
+        
+    }
+
+}
